hub: introduce package as transport agnostic communication hub.
This is my nth iteration of this concept over the years originating in the
github.com/mb0/hub package, closely coupled to websockets. The current API
is derived and reduced from package in a private project that abstracts the
transport away, to be used as a central messaging system in servers.