mig: implement project history, status and record cmds, and other cleanups
We changed the project version qualifier again to an underscore to make it
super safe as filename.

This mig package now feels right in concept. We still need to add some code
for diffing dom nodes still. ... and i need to stop writing prefix when
clearly thinking prefix, damn suffix.